Fixed typo and renamed links
#7
