+++ GoodStuff_Blazor/GoodStuff_Blazor/GoodStuff_Blazor/Components/Base/Routing.cs
@@ -1,0 +1,7 @@
+﻿namespace GoodStuff_Blazor.Components.Base;
+
+public static class Routing
+{
+    public static string SignIn => "/sign-in";
+    public static string SignUp => "/sign-in";
+}
