+++ GoodStuff_Blazor/GoodStuff_Blazor/GoodStuff_Blazor/Components/User/Pages/SignIn.razor
@@ -1,0 +1,31 @@
+﻿@page "/sign-in"
+@using GoodStuff_Blazor.Components.Base
+
+
+<div class="container col-lg-6 col-md-10 p-5 h-100">
+    <div class="h-100 shadow rounded justify-content-center col-12 align-items-center">
+        <h4 class="py-3 border-bottom">Logowanie</h4>
+        <div class="row">
+            <form class="p-2 col-6 border-right float-left" id="loginForm">
+                <div class="form-group m-2">
+                    <label>Email</label>
+                    <input type="email" name="email" id="email" class="form-control " required />
+                    <span class="text-danger " id="emailErrorMsg"></span>
+                </div>
+                <div class="form-group m-2">
+                    <label>Haslo</label>
+                    <input type="password" name="password" id="password" class="form-control " required />
+                    <span class="text-danger " id="passwordErrorMsg"></span>
+                </div>
+                <span class="btn btn-primary m-2 font-weight-bold" id="loginFormSubmit">Sign In</span>
+            </form>
+            <div class="col-6 float-left">
+                <h4>Don't have account yet?</h4>
+                <h5>It's good time to Sign Up!</h5>
+                You don't have choice anyway, without it you can't buy anything.
+                <i class="fa-regular fa-face-smile-wink"></i>
+                <a href="@Routing.SignUp"class="btn btn-primary mt-2 font-weight-bold d-block">Sign Up</a>
+            </div>
+        </div>
+    </div>
+</div>
