--- GoodStuff_Blazor/GoodStuff_Blazor/GoodStuff_Blazor/Components/User/Pages/SingUp.razor
@@ -1,76 +1,0 @@
-﻿@page "/singup"
-@using Microsoft.AspNetCore.Components.Forms
-@using Microsoft.AspNetCore.Components.Web
-
-<PageTitle>GoodStuff - Sing Up</PageTitle>
-<div class="alert alert-danger col-3 position-absolute mt-5 hide" hidden id="registrationFailMsg" role="alert">
-    <h4 class="alert-heading">O nie, cos poszlo nie tak!</h4>
-    <p>Prosze sprawdz czy formularz zostal poprawnie wypelniony.</p>
-    <hr>
-    <p class="mb-0">Mamy nadzieje ze nastepnym razem sie uda <i class="fas fa-smile-beam"></i></p>
-</div>
-
-<div class="alert alert-success col-3 position-absolute mt-5" hidden id="registrationSuccessMsg" role="alert">
-    <h4 class="alert-heading">Gratulacje, rejestracja przebiegla poprawnie!</h4>
-    <p>Na maila dostaniesz link aktywujący konto.</p>
-    <hr>
-    <p class="mb-0">Witamy w gronie GoodStuff <i class="fab fa-angellist"></i></p>
-</div>
-<div class="container p-5 h-100">
-    <div class="h-100 shadow rounded justify-content-center col-12 align-items-center">
-        <h4 class="p-2 border-bottom" style="margin-bottom: unset">Sing Up</h4>
-        <EditForm Model="@singUpModel" OnInvalidSubmit="HandleInvalidSubmit" OnValidSubmit="HandleValidSubmit" FormName="SingUpForm" method="post" class="p-2 col-6 border-right">
-            <DataAnnotationsValidator />
-            <div class="form-group m-2">
-                <label for="Name">Name</label>
-                <InputText id="Name" @bind-Value="singUpModel.Name" class="form-control" />
-                <ValidationMessage For="() => singUpModel.Name" class="text-danger" />
-            </div>
-            <div class="form-group m-2">
-                <label for="Surname">Surname</label>
-                <InputText id="Surname" @bind-Value="singUpModel.Surname" class="form-control" />
-                <ValidationMessage For="() => singUpModel.Surname" class="text-danger" />
-            </div>
-            <div class="form-group m-2">
-                <label for="Email">Email</label>
-                <InputText id="Email" @bind-Value="singUpModel.Email" class="form-control" />
-                <ValidationMessage For="() => singUpModel.Email" class="text-danger" />
-            </div>
-            <div class="form-group m-2">
-                <label for="Password">Password</label>
-                <InputText type="password" id="Password" @bind-Value="singUpModel.Password" class="form-control" />
-                <ValidationMessage For="() => singUpModel.Password" class="text-danger" />
-            </div>
-            <div class="form-group m-2">
-                <label for="ConfirmPassword">Confirm Password</label>
-                <InputText type="password" id="ConfirmPassword" @bind-Value="singUpModel.ConfirmPassword" class="form-control" />
-                <ValidationMessage For="() => singUpModel.ConfirmPassword" class="text-danger" />
-            </div>
-            <div class="p-3">
-                <h5>Accept Terms</h5>
-                <div class="form-check p-2 m-1">
-                    <InputCheckbox @bind-Value="singUpModel.AcceptTerms" class="form-check-input" id="AcceptTerms" />
-                    <label class="form-check-label" for="AcceptTerms">Oswiadczam ze nie zapoznalem sie z regulaminem strony</label>
-                    <ValidationMessage For="() => singUpModel.AcceptTerms" class="text-danger" />
-                </div>
-                <div class="form-check p-2 m-1">
-                    <InputCheckbox @bind-Value="singUpModel.AcceptEmailSpam" class="form-check-input" id="AcceptEmailSpam" />
-                    <label class="form-check-label" for="AcceptEmailSpam" >Chce otrzymywac irytujace maile o beznadziejnych promocjach</label>
-                    <ValidationMessage For="() => singUpModel.AcceptEmailSpam" class="text-danger" />
-
-                </div>
-                <div class="form-check p-2 m-1">
-                    <InputCheckbox @bind-Value="singUpModel.AcceptSellingUserData" class="form-check-input" id="AcceptSellingUserData" />
-                    <label class="form-check-label" for="AcceptSellingUserData">
-                        Zgadzam sie na to by sklep sprzedawal moje dane na prawo i lewo za dobry hajs w celu
-                        zbogacenia sie
-                    </label>
-                    <ValidationMessage For="() => singUpModel.AcceptSellingUserData" class="text-danger" />
-                </div>
-                <span class="text-danger" id="agreementsErrorMsg"></span>
-            </div>
-
-            <button type="submit" class="btn btn-primary m-2 font-weight-bold" id="registrationFormSubmit">Zaloz konto</button>
-        </EditForm>
-    </div>
-</div>
